Python-FastAPI-Archive/requirements.txt
fastapi
couchdb
python-jose[cryptography]